+++ changelog.md
@@ -24,6 +24,7 @@
 - Posts marked as stable trigger discord notifications opted in users
 - Fixed some static image links with cache busting
 - LFG thread creation bug fixes
+- Moved async tasks on game save to on commit instead of within transaction
 
 ### Breaking Changes
 - MatchAPI `tournament` field now returns a list of tournaments the game counts in instead of just the primary tournament